Slight tweaks to the specs that validate that the SUT is disposed of correctly.

+++ SpecsFor/PublishPackage.ps1
@@ -1,8 +1,16 @@
-#Remove existing packages
+Param(
+  [switch]$DoNotPush
+)#Remove existing packages
 Remove-Item *.nupkg
 #Create package
 nuget pack SpecsFor.csproj -Build -Properties Configuration=Debug
-#Push
-$PackageName = gci "*.nupkg"
-nuget push $PackageName
-Remove-Item *.nupkg
+
+if ($DoNotPush) {
+	Write-Host -ForegroundColor Yellow "Skipping 'nuget push' step!"
+}
+else {
+	#Push
+	$PackageName = gci "*.nupkg"
+	nuget push $PackageName
+	Remove-Item *.nupkg
+}
